Fix race condition which leads to test failure
The state of ServiceTaskManager is set to started only after scheduling
the MessageListeners. The message listner checks for the state of the
ServiceTaskManager and immediately terminate if it is not in the
started state. Therfore, when multiple listeners are being scheduled, the
first few tasks will still see that the active state is false and
terminate its execution bringing down the concurrent nature of the
ServiceTaskManager.