Fix thread hanging issue when deploying multiple RabbitMq listener proxies
After reviewing the Axis2 Transport code, it became evident that maintaining
a shared connection pool is unnecessary.This is because for every deployed
Listener Proxy, a new ServiceTaskManager is instantiated through RabbitMQEndpoint.
So for each Proxy a unique instance of ServiceTaskManager is available
and the Connection is hold by ServiceTaskManager.
So parallelization of connections through a pool is not needed at this
level since parrellization is happen at MessageListenerTask (inner class of ServiceTaskManager)

Here also we no longer need to implement caching or pooling
since each MessageListenerTask has its own channel created on the connection
of ServiceTaskManager and  RabbitMQ supports creating multiple channels over a
single connection. This is one of the design choices of the RabbitMQ protocol, AMQP.
Therefore, by this I changed the design to hand over connection management to reside within
the ServiceTaskManager rather than for each STM having a same shared connection pool.

Fixes: https://github.com/wso2/wso2-axis2-transports/issues/335